+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - removeMessages: method.
 - Realm database encryption.
 - OCTSubmanagerObjects: genericSettingsData property to store custom settins.
+- Adding IPv6 nodes to bootstrap.
 
 ### Changes
 - Realm updated to version 1.0.1.